Update document endpoint

+++ backend/src/data/documents/documents-dao.ts
@@ -1,4 +1,5 @@
 import { Document } from '../../models/Document';
+import mongoose from "mongoose";
 
 async function getDocuments(userId: String) {
     return Document.find({
@@ -13,8 +14,18 @@ async function createDocument(document: any) {
     return await Document.create(document);
 }
 
-async function deleteDocument(id: String) {
+async function deleteDocument(id: string) {
+    if(!mongoose.Types.ObjectId.isValid(id)) return null;
+
     return Document.findByIdAndDelete(id);
 }
 
-export { createDocument, getDocuments, deleteDocument }
+async function updateDocument(id: string, document: any) {
+    if(!mongoose.Types.ObjectId.isValid(id)) return null;
+
+    return Document.findByIdAndUpdate(id, document, {
+        new: true
+    });
+}
+
+export { createDocument, getDocuments, deleteDocument, updateDocument }
